+++ docs/beacon-transporter.fetchfn.md
@@ -1,0 +1,12 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [fetchFn](./beacon-transporter.fetchfn.md)
+
+## fetchFn variable
+
+
+<b>Signature:</b>
+
+```typescript
+fetchFn: typeof fallbackFetch
+```
